Move validators to top level namespace
