+++ src/Controller/Backend/FileManager.php
@@ -13,9 +13,9 @@ use Bolt\Filesystem\Handler\FileInterface;
 use Bolt\Filesystem\Listing;
 use Bolt\Form\FormType\FileEditType;
 use Bolt\Form\FormType\FileUploadType;
-use Bolt\Form\Validator\Constraints;
 use Bolt\Helpers\Input;
 use Bolt\Translation\Translator as Trans;
+use Bolt\Validator\Constraints;
 use Silex\ControllerCollection;
 use Symfony\Component\Form\Form;
 use Symfony\Component\Form\FormInterface;
