+++ src/Controller/Backend/General.php
@@ -6,12 +6,12 @@ use Bolt\Collection\MutableBag;
 use Bolt\Filesystem\Exception\IOException;
 use Bolt\Form\FormType\FileEditType;
 use Bolt\Form\FormType\PrefillType;
-use Bolt\Form\Validator\Constraints;
 use Bolt\Helpers\Input;
 use Bolt\Omnisearch;
 use Bolt\Requirement\BoltRequirements;
 use Bolt\Translation\TranslationFile;
 use Bolt\Translation\Translator as Trans;
+use Bolt\Validator\Constraints;
 use Doctrine\DBAL\Exception\TableNotFoundException;
 use Silex\ControllerCollection;
 use Symfony\Component\HttpFoundation\Request;
