+++ src/Validator/Constraints/FieldContains.php
@@ -1,19 +1,18 @@
 <?php
 
-namespace Bolt\Form\Validator\Constraints;
+namespace Bolt\Validator\Constraints;
 
 use Symfony\Component\Validator\Constraint;
 
 /**
- * Equivalent field value validation constraint.
+ * Field containing a value subset validation constraint.
  *
  * @author Gawain Lynch <gawain.lynch@gmail.com>
  */
-class FieldEquals extends Constraint
+class FieldContains extends Constraint
 {
-    public $message = 'This value does not equal the {{ field }} field value.';
+    public $message = 'This value does not contain the {{ field }} field value.';
     public $field;
-    public $loose = false;
     public $insensitive = false;
 
     /**
